requirements.txt
# Core
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
alembic==1.18.4
python-dotenv==1.1.2

# Media / CV (аннотатор и запекание MP4)
opencv-python-headless==4.13.0.88
pillow==11.5.0
numpy==2.3.4

# Dev
pytest==8.4.4
